fix(nix): fix nix flake in MacOs
MacOS requires a specific library `liconv` to be present during the link
stage of compiling rust programs. Though it seemed to be an input to the
flake before this change, attempting to build on MacOS yeilded this
error:

```
error: linking with `cc` failed: exit status: 1

...

= note: ld: library not found for -liconv
          clang-16: error: linker command failed with exit code 1 (use -v to see invocation)

error: could not compile `retry` (bin "retry") due to 1 previous error
```

This change updates the nix flake according to a working version
in another repository which uses `fenix` to supply rust toolchains.

It resolves the error above, and adds `fenix` as a flake input.