feat: add `--on-exit-code`
This commit refactors `retry` substantially. In doing so, tests cover
more of the system and several components with orthogonal conserns now
exist. These include:

 * Tasks
 * Deciders
 * Executors

Tasks own the actual running of the command passed to `retry`. They
generate `TaskOutcome`s which enumerate the ways a task can complete:

 * Success
 * Failure
 * Terminated
 * TimeoutExceeded

Deciders take these `TaskOutcome`s and decide if the task should be
retried or if it is `Finished`. Finished tasks cause the Executor to exit
early, while `Unfinished` tasks indicate that the Executor should retry
this task if it can.

The Executor controls how long or how many times a task should be
retried.

This refactor was instrumental in propigating `ExitStatus`s correctly
and will be invaluable when adding linear rate limiting to the Executor.